Add playRound function and edit the getChoice methods
Added the playRound function that has two parameters: one
is the human choice and the second the computer choice.
The function checks for the winner and add a point for him and
writes a message in the console.

Remove the three consts because were useless

Edited the two functions getComputerChoice and getHumanChoice
-getComputerChoice: now return PIEDRA, PAPEL or TIJERA depending
on the random number
-getHumanChoice: now return the prompt of the user. If is not PIEDRA,
PAPEL or TIJERA the function will return PIEDRA