Add three buttons and remove all the prompt interaction
Added three buttons for choice rock, paper or scissors insted of wirting
in a prompt and removed all the console.log() for display the round winner
Also added two scores: one for the cpu and the other for the player. When
you press a button the cpu choice between rock, paper and scissors randomly.
If you win the round a text appears and add one to the score and vice versa.